requirements.txt
Django==1.11.20
django-filter==2.1.0
djangorestframework==3.9.4
Markdown==3.1.1
psycopg2==2.8.2
pytz==2019.1
requests==2.22.0
sqlparse==0.3.0
